requirements.txt
pillow
google-generativeai
streamlit
watchdog